+++ scheduler/scheduler.js
@@ -111,11 +111,11 @@ module.exports = function (RED)
             if (real_topic == "set_inverted")
             {
                 real_topic = "set";
-                msg.payload = !msg.payload;
+                msg.payload = !helper.toBool(msg.payload);
             }
 
             if (real_topic == "set")
-                real_topic = (!!msg.payload) ? "enable" : "disable";
+                real_topic = helper.toBool(msg.payload) ? "enable" : "disable";
 
             switch (real_topic)
             {